REFACTOR #71 : DTO 검증로직 추가 및 댓글 목록 조회 응답에 페이징 정보를 포함하도록 수정

+++ src/main/java/com/drawandyou/drawandyou_server/global/common/response/PageInfo.java
@@ -3,10 +3,10 @@ package com.drawandyou.drawandyou_server.global.common.response;
 import org.springframework.data.domain.Page;
 
 public record PageInfo(
-        int pageNumber,
-        int pageSize,
+        long pageNumber,
+        long pageSize,
         long totalElements,
-        int totalPages,
+        long totalPages,
         boolean first,
         boolean last,
         boolean empty
